+++ host/index.jsx
@@ -110,6 +110,7 @@
   loadCmd('commands/motion.jsx');
   loadCmd('commands/follow.jsx');
   loadCmd('commands/comp.jsx');
+  loadCmd('commands/lottie.jsx');
   loadCmd('commands/fade.jsx');
   loadCmd('commands/trimpaths.jsx');
   loadCmd('commands/shapes.jsx');